Add Mutation Testing (#28)
* Add PITest plugin

* Add a couple of additional tests to kill some mutations.

* Add test coverage for TupleExpressionGenerator

Fix some copy/paste bugs in the return type interfaces.

* Expanded test suite to cover more methods

Very nice things, all the way around.

* Use Java 11 for Sonar

* Fix Gradle configuration

Apply pitest only to the fasttuple-core

+++ fasttuple-core/src/main/java/com/nickrobison/tuple/codegen/TupleExpressionGenerator.java
@@ -29,7 +29,7 @@ public class TupleExpressionGenerator extends ClassBodyEvaluator {
     }
 
     public static interface IntTupleExpression {
-        public long evaluate(FastTuple tuple);
+        public int evaluate(FastTuple tuple);
     }
 
     public static interface ShortTupleExpression {
@@ -45,11 +45,11 @@ public class TupleExpressionGenerator extends ClassBodyEvaluator {
     }
 
     public static interface FloatTupleExpression {
-        public byte evaluate(FastTuple tuple);
+        public float evaluate(FastTuple tuple);
     }
 
     public static interface DoubleTupleExpression {
-        public byte evaluate(FastTuple tuple);
+        public double evaluate(FastTuple tuple);
     }
 
     public static interface BooleanTupleExpression {
